chore: split notification generators into their own files
* previously they were all jumbled in a single file which made them
  harder to categorize
* re-exporting the classes keeps backwards-compatibility
